+++ WeatherApp/WeatherApp/Controllers/HomeViewController/DailyForecastDataSource.swift
@@ -1,55 +1,11 @@
 //
-//  ForecastDataSources.swift
+//  DailyForecastDataSource.swift
 //  WeatherApp
 //
 //  Created by Phan Quyen on 22/08/2025.
 //
 
 import UIKit
-
-protocol ForecastDataSourceDelegate: AnyObject {
-    func didSelectHourlyForecast(_ forecast: HourlyDisplayData, at index: Int)
-    func didSelectDailyForecast(_ forecast: WeeklyDisplayData, at index: Int)
-}
-
-class HourlyForecastDataSource: NSObject {
-    weak var delegate: ForecastDataSourceDelegate?
-    
-    var hourlyForecasts: [HourlyForecast] = []
-    var mockData: [HourlyDisplayData] = []
-    
-    private var displayData: [HourlyDisplayData] {
-        if !hourlyForecasts.isEmpty {
-            return hourlyForecasts.prefix(8).map { forecast in
-                HourlyDisplayData(
-                    time: forecast.timeString,
-                    temperature: forecast.temperatureString,
-                    icon: WeatherImages.imageForWeather(condition: forecast.description, iconCode: forecast.icon)
-                )
-            }
-        }
-        return mockData
-    }
-}
-
-extension HourlyForecastDataSource: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return displayData.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "ForecastCell", for: indexPath) as! ForecastCollectionCell
-        cell.configureHourly(with: displayData[indexPath.item])
-        return cell
-    }
-}
-
-extension HourlyForecastDataSource: UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        collectionView.deselectItem(at: indexPath, animated: true)
-        delegate?.didSelectHourlyForecast(displayData[indexPath.item], at: indexPath.item)
-    }
-}
 
 class DailyForecastDataSource: NSObject {
     weak var delegate: ForecastDataSourceDelegate?
@@ -72,6 +28,7 @@ class DailyForecastDataSource: NSObject {
     }
 }
 
+// MARK: - UICollectionViewDataSource
 extension DailyForecastDataSource: UICollectionViewDataSource {
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return displayData.count
@@ -84,6 +41,7 @@ extension DailyForecastDataSource: UICollectionViewDataSource {
     }
 }
 
+// MARK: - UICollectionViewDelegate & UICollectionViewDelegateFlowLayout
 extension DailyForecastDataSource: UICollectionViewDelegate {
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         collectionView.deselectItem(at: indexPath, animated: true)
